GENROUwS --> GENROU; GENSALwS --> GENSAL.
