+++ GridKit/Model/PhasorDynamics/SynchronousMachine/GENROU/Genrou.cpp
