+++ GridKit/Model/PhasorDynamics/SynchronousMachine/GENROU/GenrouEnzyme.cpp
