+++ GridKit/Model/PhasorDynamics/SynchronousMachine/GENROU/README.md
