+++ GridKit/Model/PhasorDynamics/SynchronousMachine/GENSAL/Gensal.cpp
