+++ GridKit/Model/PhasorDynamics/SynchronousMachine/GENSAL/GensalData.hpp
