+++ GridKit/Model/PhasorDynamics/SynchronousMachine/GENSAL/GensalEnzyme.cpp
