+++ examples/PhasorDynamics/Large/WECC/README.md
@@ -16,7 +16,7 @@ Model | Count
 ---|---
 [Bus](../../../../GridKit/Model/PhasorDynamics/Bus/README.md) | 243
 [Branch](../../../../GridKit/Model/PhasorDynamics/Branch/README.md) | 447
-[GENROU](../../../../GridKit/Model/PhasorDynamics/SynchronousMachine/GENROUwS/README.md) | 140
+[GENROU](../../../../GridKit/Model/PhasorDynamics/SynchronousMachine/GENROU/README.md) | 140
 [TGOV1](../../../../GridKit/Model/PhasorDynamics/Governor/Tgov1/README.md) | 103
 [SEXS-PTI](../../../../GridKit/Model/PhasorDynamics/Exciter/SEXS-PTI/README.md) | 103
 [IEEEST](../../../../GridKit/Model/PhasorDynamics/Stabilizer/IEEEST/README.md)  | 10
